Explicitly check that report object loggers don't propagate
This is the desired behavior, so we should check this.

We strass this, becasue if we want to use `pytest` to capture logs, this will
only work with loggers that propagate. We want to explicitly state that we
don't want loggers to propagate and that we will need to work around this
behavior in the tests, not by modifying the production code.