Use `pytest.caplog` to test logging
We need to add a bit of code to inject the `caplog` fixture, but this will go
away once we remove the `TestCase` class.

`caplog` will reset logging levels after each test, so no need for context
handlers.

`caplog` only captures log messages through the root logger, so we unfortunately
have to work around this by making loggers propagate.

We had to change the checks to check log records instead of messages. This isn't
such a bad thing, because messages will depend on the format that was
configured and hence lead to brittle tests.

It should be pretty easy to see what steps were required to use `caplog`,
because we split the tests beforehand. (This resulted in quite a bit of
duplication, so we should think of refactoring to fix this.)