Extract functions for adjacency checks
Flake8 and PEP8 disagree when it comes to breaking expressions. PEP8
recommends breaking before a binary operator (as recommended by Donald
Knuth), whereas Flake8 wants breaks to happen after a binary operator
(as PEP8 used to mandate).

Extracting some functions for the expressions avoids the issue. It also
improves readability a bit more, because it gives names to the code
snippets.