Refactored uvm_export_base and uvm_port_base
Fixed problems in the pyuvm classes where you could call
connect on an export which would not return an error and
would not work.  Now uvm_export_base does not have connect
and the check_export() method uses a list of function names
and hassattr() to check that the export implemented the
needed methods.

Issue #35