Now start_item actually waits for driver
This release fixed a bug in the sequence system in which the start_item() coroutine didn't actually wait for the driver to get the item. While it guaranteed order it did not guarantee timing so it was not possible to add late stimulus between start_item and finish_item.

Now tests for this behavior and implements it.